feat(kener-status): create initial changeset for plugin
Create changeset for first release of Kener Status plugin with:
- Add changeset configuration
- Set minor version change for initial release
- Update package metadata
