+++ workspaces/kener/kener-status/.changeset/hungry-houses-sleep.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-kener-status': minor
+---
+
+This plugin provides integration with Kener to display uptime and status badges in Backstage system pages.
