+++ workspaces/kener/kener-status/package.json
@@ -21,11 +21,13 @@
     "test": "backstage-cli package test",
     "clean": "backstage-cli package clean",
     "prepack": "backstage-cli package prepack",
-    "postpack": "backstage-cli package postpack"
+    "postpack": "backstage-cli package postpack",
+    "changeset": "changeset"
   },
   "dependencies": {
     "@backstage/core-components": "^0.16.3",
     "@backstage/core-plugin-api": "^1.10.3",
+    "@backstage/plugin-catalog-react": "^1.15.1",
     "@backstage/theme": "^0.6.3",
     "@material-ui/core": "^4.9.13",
     "@material-ui/icons": "^4.9.1",
@@ -40,6 +42,7 @@
     "@backstage/core-app-api": "^1.15.4",
     "@backstage/dev-utils": "^1.1.6",
     "@backstage/test-utils": "^1.7.4",
+    "@changesets/cli": "^2.27.1",
     "@testing-library/jest-dom": "^6.0.0",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.0.0",
