+++ workspaces/kener/kener-status/src/index.ts
@@ -1,3 +1,7 @@
 // plugins/kener-status/src/index.ts
-export { kenerStatusPlugin, rootRouteRef, EntityKenerStatusCard } from './plugin';
+export {
+  kenerStatusPlugin,
+  rootRouteRef,
+  EntityKenerStatusCard,
+} from './plugin';
 export { KenerStatusPage } from './components/KenerStatusPage';
